# projects/notification-management-api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing email, SMS, and push notifications",
        "title": "Notification Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/sms/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List SMS notifications with optional date and status filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "List SMS",
                "parameters": [
                    {
                        "enum": [
                            "PENDING",
                            "SENT"
                        ],
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 50,
                        "minimum": 0,
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "create_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sms.GetAllSmsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Queue a single SMS notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "Send SMS",
                "parameters": [
                    {
                        "description": "Send SMS Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sms.SendSmsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/sms/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Queue multiple SMS notifications (max 1000)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "Send batch SMS",
                "parameters": [
                    {
                        "description": "Send Batch SMS Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sms.SendBatchSmsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/sms/cancel": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cancel SMS notifications matching the given status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "Cancel SMS",
                "parameters": [
                    {
                        "enum": [
                            "PENDING"
                        ],
                        "type": "string",
                        "description": "Status to cancel",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/sms/status": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the status of a single SMS notification",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "Get SMS status by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SMS ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sms.GetSmsStatusByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/sms/status/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the status of multiple SMS notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "Get SMS statuses by IDs",
                "parameters": [
                    {
                        "description": "Batch Status Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sms.GetStatusByBatchIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sms.GetStatusByBatchIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "error_handling.Errors": {
            "type": "object",
            "properties": {
                "CorrelationID": {
                    "type": "string"
                },
                "Err": {},
                "ErrorType": {
                    "type": "string"
                },
                "Message": {
                    "type": "string"
                },
                "Parameter": {
                    "type": "string"
                },
                "StackTrace": {
                    "type": "string"
                },
                "StatusCode": {
                    "type": "integer"
                }
            }
        },
...
        "sms.GetAllSmsResponse": {
            "type": "object",
            "required": [
                "page",
                "page_size"
            ],
            "properties": {
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "page_size": {
                    "type": "integer",
                    "maximum": 50,
                    "minimum": 0
                },
                "smsList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sms.Sms"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "sms.GetSmsStatusByIDResponse": {
            "type": "object",
            "properties": {
                "sms_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "sms.GetStatusByBatchIDRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "sms.GetStatusByBatchIDResponse": {
            "type": "object",
            "properties": {
                "smsList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sms.GetSmsStatusByIDResponse"
                    }
                }
            }
        },
        "sms.SendBatchSms": {
            "type": "object",
            "required": [
                "content",
                "phone_number",
                "priority",
                "sender",
                "type"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 160,
                    "minLength": 1
                },
                "phone_number": {
                    "type": "string"
                },
                "priority": {
                    "type": "string",
                    "enum": [
                        "LOW",
                        "MEDIUM",
                        "HIGH"
                    ]
                },
                "scheduled_at": {
                    "type": "string"
                },
                "sender": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 2
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "sms.SendBatchSmsRequest": {
            "type": "object",
            "properties": {
                "sms": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/sms.SendBatchSms"
                    }
                }
            }
        },
        "sms.SendSmsRequest": {
            "type": "object",
            "required": [
                "content",
                "phone_number",
                "priority",
                "sender",
                "type"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 160,
                    "minLength": 1
                },
                "phone_number": {
                    "type": "string"
                },
                "priority": {
                    "type": "string",
                    "enum": [
                        "LOW",
                        "MEDIUM",
                        "HIGH"
                    ]
                },
                "scheduled_at": {
                    "type": "string"
                },
                "sender": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 2
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "sms.Sms": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "idempotency_key": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "integer"
                },
                "sender": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}